工商联统计分析页面优化

+++ ace/ace-common-web/src/main/java/com/huacainfo/ace/common/web/tools/WebUtils.java
@@ -120,13 +120,15 @@ public class WebUtils {
         AudioAttributes audio = new AudioAttributes();
         Encoder encoder = new Encoder();
         audio.setCodec("libmp3lame");
-        audio.setBitRate(new Integer(86000));
+        audio.setBitRate(new Integer(64000));
         audio.setChannels(new Integer(1));
         audio.setSamplingRate(new Integer(22050));
+
+
         VideoAttributes video = new VideoAttributes();
         video.setCodec("libxvid");// 转MP4
-        video.setBitRate(new Integer(860000));
-        video.setFrameRate(new Integer(15));
+        video.setBitRate(new Integer(500000));
+        video.setFrameRate(new Integer(1));
         EncodingAttributes attrs = new EncodingAttributes();
         attrs.setFormat("mp4");
         attrs.setAudioAttributes(audio);
